+++ logme/exceptions.py
@@ -14,5 +14,9 @@ class InvalidConfig(Exception):
     """Used when invalid configuration is passed"""
 
 
+class InvalidColorConfig(Exception):
+    """Used when invalid color configuration is passed"""
+
+
 class DuplicatedHandler(Exception):
     """Used when an identical handler is being added on to a logger"""